requirements.txt
colouredlogs==10.0
msgpack==0.5.6
prompt-toolkit==2.0.9
pyzmq==18.0.2
six==1.12.0
wcwidth==0.1.7
